pyproject.toml
[project]
name = "mcp-pandoc"
version = "0.8.1"
description = "MCP to interface with pandoc to convert files to different formats with enhanced features like Mermaid diagram conversion and defaults file support."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
 "mcp>=1.2.1",
 "pyyaml>=6.0.2",
 "pandoc>=2.4",
 "pypandoc>=1.14",
 "pandocfilters>=1.5.0",
 "panflute>=2.3.1",
]
[[project.authors]]
name = "Vivek Vellaiyappan Surulimuthu"
email = "vivekvellaiyappans@gmail.com"

[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest-asyncio>=1.0.0",
    "pytest>=8.4.1",
    "ruff>=0.12.11",
    "jsonschema>=4.25.1",
    "yamllint>=1.37.1",
    "pre-commit>=4.3.0",
]

[project.scripts]
mcp-pandoc = "mcp_pandoc:main"
[tool.ruff]
line-length = 120
exclude = ["tests/*"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "D", "N", "S"]
ignore = []

[tool.yamllint]
config-file = ".yamllint.yaml"
